minor fixes

+++ server/game_logic/characters/enemies.cpp
@@ -14,12 +14,12 @@
 #define LIZARD_GOON_HEIGHT 50
 #define LIZARD_GOON_SPEED 75
 
-MadHatter::MadHatter(size_t id, int x, int y):
+MadHatter::MadHatter(uint16_t id, int x, int y):
         Enemy(id, MAD_HATTER, MAD_HATTER_ATTACK_DAMAGE, MAD_HATTER_HEALTH,
               MAD_HATTER_REVIVE_COOLDOWN, x, y, MAD_HATTER_WIDTH, MAD_HATTER_HEIGHT,
               MAD_HATTER_SPEED) {}
 
-LizardGoon::LizardGoon(size_t id, int x, int y):
+LizardGoon::LizardGoon(uint16_t id, int x, int y):
         Enemy(id, LIZARD_GOON, LIZARD_GOON_ATTACK_DAMAGE, LIZARD_GOON_HEALTH,
               LIZARD_GOON_REVIVE_COUNTDOWN, x, y, LIZARD_GOON_WIDTH, LIZARD_GOON_HEIGHT,
               LIZARD_GOON_SPEED) {}
