+++ server/game_logic/characters/enemies.h
@@ -2,10 +2,10 @@
 
 class MadHatter: public Enemy {
 public:
-    MadHatter(size_t id, int x, int y);
+    MadHatter(uint16_t id, int x, int y);
 };
 
 class LizardGoon: public Enemy {
 public:
-    LizardGoon(size_t id, int x, int y);
+    LizardGoon(uint16_t id, int x, int y);
 };
