+++ server/game_logic/characters/player.cpp
@@ -1,7 +1,7 @@
 #include "player.h"
 
 
-Player::Player(size_t id, std::string name, const character_t& character, int x, int y,
+Player::Player(uint16_t id, std::string name, const character_t& character, int x, int y,
                CollisionManager& collision_manager):
         CharacterBody(id, character, x, y, PLAYER_WIDTH, PLAYER_HEIGHT,
                       Vector2D(NONE, MAX_FALL_SPEED), MAX_HEALTH, STATE_IDLE_RIGHT,
