+++ server/game_logic/characters/players.cpp
@@ -1,10 +1,10 @@
 #include "players.h"
 
-Jazz::Jazz(size_t id, const std::string& name, int x, int y, CollisionManager& collision_manager):
+Jazz::Jazz(uint16_t id, const std::string& name, int x, int y, CollisionManager& collision_manager):
         Player(id, name, JAZZ_CHARACTER, x, y, collision_manager) {}
 
-Spaz::Spaz(size_t id, const std::string& name, int x, int y, CollisionManager& collision_manager):
+Spaz::Spaz(uint16_t id, const std::string& name, int x, int y, CollisionManager& collision_manager):
         Player(id, name, SPAZ_CHARACTER, x, y, collision_manager) {}
 
-Lori::Lori(size_t id, const std::string& name, int x, int y, CollisionManager& collision_manager):
+Lori::Lori(uint16_t id, const std::string& name, int x, int y, CollisionManager& collision_manager):
         Player(id, name, LORI_CHAARCTER, x, y, collision_manager) {}
