+++ server/game_logic/characters/players.h
@@ -4,15 +4,15 @@
 
 class Jazz: public Player {
 public:
-    Jazz(size_t id, const std::string& name, int x, int y, CollisionManager& collision_manager);
+    Jazz(uint16_t id, const std::string& name, int x, int y, CollisionManager& collision_manager);
 };
 
 class Spaz: public Player {
 public:
-    Spaz(size_t id, const std::string& name, int x, int y, CollisionManager& collision_manager);
+    Spaz(uint16_t id, const std::string& name, int x, int y, CollisionManager& collision_manager);
 };
 
 class Lori: public Player {
 public:
-    Lori(size_t id, const std::string& name, int x, int y, CollisionManager& collision_manager);
+    Lori(uint16_t id, const std::string& name, int x, int y, CollisionManager& collision_manager);
 };
