+++ server/game_logic/weapons/guns.cpp
@@ -11,8 +11,8 @@ Constructor for guns:
         - collision_manager: collision manager of the game
 
         ----- personal config -----
-        -size_t ammo,
-        -size_t max_ammo,
+        -uint8_t ammo,
+        -uint8_t max_ammo,
         -int weapon_damage,
         -int shoot_rate
 */
@@ -23,7 +23,7 @@ Constructor for guns:
  * A lot of Shoot rate
  * Little Bullet speed
  */
-GunOne::GunOne(size_t weapon_id, Player& player_owner, CollisionManager& collision_manager):
+GunOne::GunOne(uint8_t weapon_id, Player& player_owner, CollisionManager& collision_manager):
         Weapon(weapon_id, player_owner, collision_manager, 5, 200, 60, 20) {}
 
 /*
@@ -32,7 +32,7 @@ GunOne::GunOne(size_t weapon_id, Player& player_owner, CollisionManager& collisi
  * Little Shoot rate
  * A lot of Bullet speed
  */
-GunTwo::GunTwo(size_t weapon_id, Player& player_owner, CollisionManager& collision_manager):
+GunTwo::GunTwo(uint8_t weapon_id, Player& player_owner, CollisionManager& collision_manager):
         Weapon(weapon_id, player_owner, collision_manager, 25, 25, 10, 20) {}
 
 /*
@@ -41,7 +41,7 @@ GunTwo::GunTwo(size_t weapon_id, Player& player_owner, CollisionManager& collisi
  * A lot of Shoot rate
  * A lot of Bullet speed
  */
-GunThree::GunThree(size_t weapon_id, Player& player_owner, CollisionManager& collision_manager):
+GunThree::GunThree(uint8_t weapon_id, Player& player_owner, CollisionManager& collision_manager):
         Weapon(weapon_id, player_owner, collision_manager, 50, 10, 60, 20) {}
 
 /*
@@ -51,7 +51,8 @@ GunThree::GunThree(size_t weapon_id, Player& player_owner, CollisionManager& col
  * regular shoot rate
  */
 
-DefaultGun::DefaultGun(size_t weapon_id, Player& player_owner, CollisionManager& collision_manager):
+DefaultGun::DefaultGun(uint8_t weapon_id, Player& player_owner,
+                       CollisionManager& collision_manager):
         Weapon(weapon_id, player_owner, collision_manager, 10, 0, 60, 20) {}
 
 void DefaultGun::shoot() {
@@ -59,9 +60,9 @@ void DefaultGun::shoot() {
     if (!can_shoot) {
         return;
     }
+
     shoot_rate_counter = 0;
     uint64_t bullet_id = create_bullet_id();
-    // TODOD --> get list of bullets reference from player
     auto bullet = std::make_shared<Bullet>(bullet_id, weapon_id, player_owner, weapon_damage);
     collision_manager.track_dynamic_body(bullet);
 }
