+++ server/game_logic/weapons/guns.h
@@ -6,22 +6,22 @@
 
 class GunOne: public Weapon {
 public:
-    GunOne(size_t weapon_id, Player& player_owner, CollisionManager& collision_manager);
+    GunOne(uint8_t weapon_id, Player& player_owner, CollisionManager& collision_manager);
 };
 
 class GunTwo: public Weapon {
 public:
-    GunTwo(size_t weapon_id, Player& player_owner, CollisionManager& collision_manager);
+    GunTwo(uint8_t weapon_id, Player& player_owner, CollisionManager& collision_manager);
 };
 
 class GunThree: public Weapon {
 public:
-    GunThree(size_t weapon_id, Player& player_owner, CollisionManager& collision_manager);
+    GunThree(uint8_t weapon_id, Player& player_owner, CollisionManager& collision_manager);
 };
 
 class DefaultGun: public Weapon {
 public:
-    DefaultGun(size_t weapon_id, Player& player_owner, CollisionManager& collision_manager);
+    DefaultGun(uint8_t weapon_id, Player& player_owner, CollisionManager& collision_manager);
     void shoot() override;
 };
